+++ src/Vctoon.Application.Contracts/Libraries/Dtos/ImageFileDto.cs
@@ -1,4 +1,3 @@
-using System;
 using Volo.Abp.Application.Dtos;
 
 namespace Vctoon.Libraries.Dtos;
@@ -12,11 +11,11 @@ public class ImageFileDto : EntityDto<Guid>
 
     public string Extension { get; set; }
 
-    public uint Size { get; set; }
+    public long Size { get; set; }
 
-    public uint Width { get; set; }
+    public long Width { get; set; }
 
-    public uint Height { get; set; }
+    public long Height { get; set; }
 
     public Guid? LibraryPathId { get; set; }
