+++ src/Vctoon.Application.Contracts/Libraries/Dtos/ImageFileGetListInput.cs
@@ -11,11 +11,11 @@ public class ImageFileGetListInput : PagedAndSortedResultRequestDto
 
     public string Extension { get; set; }
 
-    public uint? Size { get; set; }
+    public long? Size { get; set; }
 
-    public uint? Width { get; set; }
+    public long? Width { get; set; }
 
-    public uint? Height { get; set; }
+    public long? Height { get; set; }
 
     public Guid? LibraryPathId { get; set; }
